Fix _generateName() to prefer the name (duh)
Long, long ago the decision was made to prefer the displayedText over the
accessible name. To be honest, I'm not sure why. I've subsequently not
changed that for fear of breaking something unexpected. But bugs keep
coming up because of this old decision. We'll fix them based on reports
before this is included in stable.