Fix bug causing us to not always use the system voice
The speech generator doesn't use the system voice when it's not been
"established." The preferences GUI wasn't setting that key. In contrast,
the message presentation doesn't check for that. Having fixed the bug
with "established", the system voice will now be used for everything it
was meant (coded) to. If that produces unexpected results, we can tweak
the "meant (coded) to" as needed.