Don't insert a pause when punctuation is set to all
To accomplish breaking up speech into chunks separated by pauses, we
append a period because that seems to work regardless of synthesizer.
Unfortunately, when punctuation is set to all, the synthesizer speaks
that period.

We can probably improve our logic to break things up in another way
and/or request API from speech-dispatcher for pause insertion. In the
meantime, prevent the period from being spoken.