Prevent skip links which appear above banners from being treated on same line
When a site has a skip link which appears above a banner, it can shove the
elements in the banner down. Unfortunately, this doesn't cause the user
agents to give us the updated bounding boxes. As a result, it seems like
the skip link is on the same line as the repositioned banner. In order to
try to catch this condition, check links for banner ancestors and treat
them as if they are on a different line if one link is in a banner and the
other is not.