Only treat block list descendants on different lines if they are not links
The goal of treating block list descendants on different lines is to keep
text blocks from being presented on the same line when functionally they
are not (e.g. a heading on the same line as the start of a list). If
we have a set of links in a list, even if the list items report display:
block, we want to let the bounding boxes determine whether or not to
treat them as if they are on the same line.