Adjust when we clear the active window and focus in event manager
If we got a window:deactivate event and the event source was defunct,
we would automatically clear the active window and locus of focus in
the event manager without checking to see if the event source was the
active window. There's a very small chance it won't be, so check first.

In addition, we weren't handling the case where an app fails to emit
window:deactivate, but goes straight to window:destroy. So also check
and handle that condition.