Also clear active script when we get an event for the defunct active window
When a window is deactivated, but not defunct/dead, the default script
clears the locusOfFocus and the active window AND the active script. But
the event manager was only clearing the first two when handling the window
going defunct/dead on us. This can cause us to have a stale/bogus active
script. Therefore, follow the same behavior as the default script.

Note: Given a badly behaved window manager / switcher, some recovery work
will be needed to update our state. This will be done in a subsequent
commit as it's arguably a separate problem and one which exists even if
we don't change the event manager.