Update our state when input event has to search for the active window
When we didn't have a saved active window when an input event came in,
we corrected by figuring out the active window. But we didn't save the
state based on our findings. As a result, we sometimes did this work
multiple times in a row until an app gave us an appropriate event.