Improve presentation of caret moved events to end of contenteditable children
When navigating in contenteditable content, moving just past a child
element can result in our getting a caret-moved event for the end of that
child, but not one for what is functionally the new caret position. We were
already handling that for contenteditable links, so just remove the link
restriction.