Don't look to the next element when line nav moves us to the end
When navigating in contenteditable content, moving just past a child
element can result in our getting a caret-moved event for the end of that
child, but not one for what is functionally the new caret position. Our
handling for this situation is producing incorrect results when we have
wound up at the end of an element as a consequence of line navigation.
Therefore don't adjust the offset when navigating by line.