Add more logic to try to find valid selection container
We cannot count there not being bogus items in the accessibility tree,
such as an unexpected role section standing in between a tree item and
the tree container. Therefore ascend the hierarchy further and check
roles to ensure we don't return in irrelevant ancestor which happens
to implement the selection interface.