Gecko: Use getAccessibleAtPoint() to get the accessible at coordinates
Don't do manual checks and rely on Gecko for performing the lookup, as
it allows us to get accurate results.

Although it might be unexpected, Gecko's getAccessibleAtPoint() returns
the accessible child that is in the path leading to the one actually at
the coordinates, even if it's not itself containing them.

Apart from the recursion in _accessibleAtPoint(), the logic is mostly
stolen from Chromium script.

https://gitlab.gnome.org/GNOME/orca/issues/83