Remove redundant method
The default script utility getContainingDocument() is doing essentially
the same work that the web script utility getDocumentForObject is doing.
The only difference is that the latter checks if something is a document
first, which callers of the former were in some cases doing. So move the
latter to the default script and remove the former.