Thunderbird: Ignore busy-changed events when in focus mode
Thunderbird's settings UI now consists of web content, in which choosing
an option in the tree on the left causes us to get events which look like
a new document or message was loaded. If we ignore these events in focus
mode, the user's location will not be incorrectly changed to the document.