Gecko: Fix finding the text elements containing children
orca.script_utilities.descendantAtPoint() will not consider the
passed-in node as the final target if it can usually have children, but
will accept such nodes if they are non-empty text nodes it found
through their parent.

As Gecko's implementation now first skips directly to what Gecko thinks
is the element at point and chains up to the parent implementation with
the found element to handle cases where we don't actually want to
present it, parent implementation that also rejects the aforementioned
nodes.

It works in Firefox because of a side effect of it erroneously
returning text nodes through getAccessibleAtPoint() (although it
normally doesn't expose them), we return the parent node of that text
node directly, skipping the code that would discard that parent node.

But in Thunderbird those text nodes are not exposed, so we need to
properly handle this case instead of relying on the unexpected text
nodes.