Fix window list the second time mouse review gets enabled
On initial Wnck calls the signals we connect to get emitted for the
initial values, but in case we got disabled and re-enabled we have to
get the initial values manually.

This could in theory be avoided by shutting down Wnck when we get
disabled, and then the next activation would behave like the initial
one, but Wnck.shutdown() is documented as being hacky, and seem too
buggy (spews lots of warnings that are not related to our code).