Verbalize punctuation for the new ARIA code role
According to the ARIA spec, for the code role screen readers should
prefer full punctuation verbosity to ensure common symbols (e.g. "-")
are spoken. In order to balance this with user preference, we'll only
do this when the user has their punctuation level set to some or most,
which suggests they are interested in relevant punctuation but have not
explicitly turned over this work to their speech synthesizer.