Speak "alert" role before its name rather than after
We are seeing some alerts whose accessible name is the start of the
alert message. This results in an awkward presentation with the role
interrupting the message. Since we cannot control/fix all the alerts
out there, reversing the order of presentation seems like the sanest
approach. Plus doing so alerts (as it were) the user to the nature of
the message about to be spoken.