Ignore caret-moved events from outside the active grid in focus mode
If the user is arrowing within a grid in focus mode we should not be
getting caret-moved events from outside the grid. And if we are, we
should not present them nor update our state or mode. We are seeing
this sort of thing happen in (at least) Chromium with (at least)
Gmail when the user attempts to arrow past the last column.