Don't iterate through all children of very complex SVGs
If an SVG had children, even though it lacked any other presentable
information, we treated it as potentially worthy of presentation.
Then we descended the children looking for objects worthy of
presentation. But if an SVG is sufficiently complex, it can have
thousands of children, and any recursive method used to iterate
through them can result in a crash. Therefore, check the first 50
children, and if they also lack presentable information, treat the
SVG as not being worthy of presentation.