Only present the status bar if it is showing and visible
Some applications might have multiple status bars, only one of which
is on the screen and should be presented to the user. On-screen objects
are expected to have both the showing and the visible states present.
Therefore check for this condition before assuming we have located the
correct status bar.