Update the active script upon launch based on focused object
When Orca is launched we were setting the active script based on the
active window, but not adjusting it for the focused object. Usually
the focused object script is the same as the script for the active
window. But this is not always the case for embedded widgets such as
terminals.