Don't announce returning to browse mode when page is loading
Often we get focus events when a page is loading that can cause us to
switch to browse mode. This can be followed by another event which
causes us to immediately switch back to focus mode. Hearing Orca say
"browse mode focus mode" is not especially useful.

We'll continue to announce switching to focus mode even when a page is
loading because certain web apps emit focus events for widgets while
the page is still loading.