Improve heuristic to determine when switching back to browse mode for link
Orca was changed to not automatically switch back to browse mode when
the user was in focus mode and then focus moved to a link. This was done
to prevent Orca from returning to browse mode for web apps where the
author was using links for interactive lists (e.g. in Gmail). But this
change had the side effect of Orca not returning to browse mode when
the user is Tabbing amongst widgets and links. Therefore, for now limit
remaining in focus mode when an unmodified arrow was used from within
focus mode.