Add ability to infer labels from layout tables exposed as sections
Orca has label inference code to work around authors who for whatever
reason fail to properly label their form fields. This label inference
code also takes into account layout tables, but that aspect of the
solution was failing to work in Chromium because Chromium does not
expose layout tables as tables (a good thing on Chromium's part, but
a bad thing for the inference code which was expecting to use the
table interface based on what Firefox does). Therefore, add handling
for layout tables exposed as sections to Orca's label inference code.