Reduce chattiness from page tab selection events following focus
If we get an event from a page tab selection, and if the selected page
tab has the same name as the currently focused object, there's a healthy
chance that the page tab selection event came after the focus event. In
this instance, updating the current focus and presenting the change leads
to unwanted chattiness.

While this problem and fix resulted from a VSCode issue, this scenario
could easily happen in other applications. So put the handling in the
default script.