Remove attempt to filter out duplicate names calculated by content
A workaround for a chattiness problem was introduced a couple of years
ago to deal with duplicate names resulting from name being calculated
from contents as per AccName. And we've been modifying the workaround
ever since. This sort of thing should be addressed in the AccName script,
user agents, and authoring practices rather than hacked around here.