Handle caret-moved event emitted before focus events in newly-loaded page
Sometimes we get a caret-moved event for a fragment before the page
has finished loading and/or the document claimed focus. In this case,
the eventIsFromLocusOfFocusDocument() test fails because the locusOfFocus
document is a zombie. Therefore, if we detect this condition, compare
the event.source document with the active document based on the embeds
relation.