Chromium: Fix issue causing us to present the previous line in contenteditable
We don't want to adjust the offset for literal newline characters in
contenteditable content. Chromium gives us the right information, so
we should trust it. In the case of Gecko, we already sanity check
everything it gives us for text at offset because its text implementation
is broken. Any adjustments which are not already being made for it should
be made in _getTextAtOffset().