Braille: Add additional attempts to recover from broken apps/toolkits
We're already doing a lot of work to determine what should be the first
visible object on the braille display (aka the "focused region"). This
work is needed because certain apps and toolkits seem to like to destroy
and recreate objects. Calling isSameObject() solves much of that problem,
but apparently it can still fail. Therefore if all of our hard work still
fails to find a match and we have a single object with the same name and
role, prefer it over what is likely an ancestor object so users do not
have to pan their display to the right to get to what is being spoken.

Also add some more debugging to get to the bottom of app/toolkit bugs
which I cannot reproduce myself.