Braille: Ensure focused region starts the display when panToCursor
When we update the braille for a particular object, braille.refresh()
would sometimes show the object at the left of the display, but other
times not. The difference was whether the object could fully fit on
the display. Thus when navigating to a menu with long names, the display
started with the menu name. But if the name was short, the application
name (or the end of it) might show as a prefix to the menu name. This
inconsistency seems bad. So stop it.