Chromium: Use caret-moved event to recover from object destruction
Sometimes when objects are destroyed as a side effect of caret navigation,
Chromium gives us a caret-moved event for the new location. Using this
event allows us to gracefully and efficiently recover from the destruction.

Solves (at least) the problem of Orca's caret-navigation failing after
Github's actions drop-down is accessed in Browse mode and then collapses.