Only update context at end of contenteditable element during char nav
We were updating the caret content when the caret moved to the end of
a contenteditable element unless the movement resulted from line nav
or line boundary nav. The idea was that if char nav or word nav were
used, the behavior would be wanted. But some editors give us caret-moved
events for the end of an object as a consequence of inserting or deleting
text. In those instances, updating our context can cause the default
script to not present the caret moved event.

More research should be done to determine if this update should be done
for word navigation.