Ensure we present functionally-empty entries from web apps
Entries which are functionally empty (such as the message entry in Riot)
but have a textblock element serving as a placeholder were being treated
as not being a valid caret context. The reason why was that we should have
been treating this entry as a whole rather than descending its children.

Once the above issue was fixed, we were filtering out the entry for
presentation because it was being treated as a potentially more complex
contenteditable with no presentable string / content of its own. So don't
treat entries whose sole child is a functional placeholder as such a beast.