braille: remove fallback on VT 7
X or Wayland is less and less actually running on VT 7. libbrlapi has been
enhanced to ask logind which VT we are running on even when WINDOWPATH or
XDG_VTNR is not set, so avoid interfering with that by passing a
hardcoded VT number.

At worse, users can use XDG_VTNR to override the VT number determined by
libbrlapi.