Work around Qt failing to use correct role and/or state
Checkable menu items should have STATE_CHECKABLE and/or the role of
ROLE_CHECK_MENU_ITEM. Apparently this is not always the case for Qt
apps. This workaround should cause Orca to announce the checked state
when present, but not announce "not checked." In order for the latter
to occur, Qt apps should fix the aforementioned bug.