Don't return early when we get caret-moved event from empty editable objects
From user debugging output, it appears that Thunderbird/Gecko will emit
caret-moved events for objects which have zero characters. Arguably this
doesn't make sense: If the caret supposedly moved to the first character
in an element, there is a first character in the element and thus the
element should have a character count. But Gecko's text implementation is
pretty broken, so try to be more forgiving.