Don't bother checking input and textarea elements as complex contenteditable
We want to check editable elements for descendants so we handle rich-text
editors correctly. There's no point in doing this for native input and
textarea elements. Granted, checking for descendants in childless elements
should be super quick. But we're seeing some weird performance problems
when doing this check for giant textareas in Chromium.