Implement initial recovery from the current focused object being removed
If we get a child-removed event for the removal of the current focus,
and if that removal immediately followed line nav, attempt to locate
the nearest object to the removed child based on the direction of
navigation. If we find such an object, move there and present the line.

Note: The goal of this recovery is not perfection; it's to improve
unpredictable jumping around when the web author destroys the object
we're in.

This seems to improve the situation when we wander into Riot's Message
Actions toolbar in Chromium and then that toolbar gets destroyed while
we're in it. (The issue with our getting stuck in that same toolbar is
a different issue which still needs to be solved.)