Prevent children which cannot have context being returned as context
find{Next,Previous}CaretInOrder() was returning children which should
be treated as a whole as valid caret context without checking if
said child could actually serve as a caret context. Objects which cannot
serve as a caret context include objects we which to ignore because they
are dead, zombies, empty, hidden, etc., etc.

As a consequence of this bug, we were sometimes wandering into hidden
objects (not good), presenting them (worse), and then getting stuck in
them (doubleplusbad).

This failure seems to be responsible for our getting stuck in Riot in
(at least) Chromium.