In moving the review cursor to the start of the braille
display, obtain an offset relative to the Accessible.

The word finding logic in flat review requires an offset relative to the
Accessible, not relative to the Zone. Orca was failing as a result to
pan the braille display correctly on the second and subsequent lines of
text (e.g., in GNOME-Terminal).