Improve heuristic to identify elements serving as fake placeholder for text
We want to treat fake placeholders for fake inputs like placeholders and
not confuse them for editable content in a contenteditable. The existing
heuristic was working for Firefox, but not for Chromium due to extra divs
being included in the accessibility tree and also text leaf nodes being
incorrectly treated as editable text elements.