Ensure we announce role and content of elements with ARIA's separator role
In the case of Chromium, we were announcing the role, but not checking
for text content; in the case of Firefox, we were finding a child of the
separator with text but not then presenting the ancestor separator. This
fix causes us to handle both cases.

Also, don't present values for non-focused separators. Non-focusable
separators are not supposed to implement the value interface. But if an
author slaps a tabindex value of -1 on them, then user agents seem to
treat them as focusable and expose the value interface on them, complete
with default values which are likely useless noise given the object is
not serving as a widget.