Exclude ancestors when getting the labels for an object
In some applications/toolkits (e.g. Qt), the group/panel containing
the widgets claims to be labelling its child widgets. If we don't
filter this out, we are in danger of repeating the group label or
even replacing the widget's label with the group label. We already
present the ancestor group as part of the context of widgets, so we
can eliminate it here.