Look for the deepest presentable items when generating the status bar
_generateStatusBar was only looking at and presenting immediate children,
and then prefering the displayed text of that item. This approach can
cause us to miss certain items, as well as to incorrectly omit the role
on interactive items, especially in web applications.

Because status bars, even on the web, shouldn't have that many descendant
objects, dive the tree locating presentable descendants.