In calculating the braille cursor offset in flat review,
include any white space at the beginning of the zone.

If this space is not taken into account, Orca can fail to pan the
braille display to the right on indented lines. In particular, if word wrap is enabled,
the incorrectly calculated cursor offset can remain within the current
range, resulting in the display's not being panned at all.