Clear braille-related data structures on moving to the next line
in flat review.

Braille-related variables were not initialized upon moving to the next
line, resulting in old values, notably of viewport[0] being used. The
effect was that a "pan right" operation moving to the next line could be
treated as a "pan
left"; if word wrap were in effect, the ranges from the current line
would be looked up using the offset location from the previous line,
leading to a failure to pan the display.