Clarify and simplify braille right panning code.
The call to reviewNextCharacter() implicitly updates and refreshes the
braille display to show the next line. From the original logic of
panBrailleRight(), it misleadingly appeared as though the subsequent
call to updateBrailleReview was responsible for this work. Also, this
later call was redundant. We eliminate the redundant call and clarify
the logic of this code to avoid causing confusion to future developers,
with a slight efficiency gain.