Fix cell coordinate retrieval on GTK
Cells need not directly be inside a table, but can have other cells as
parents: cells can contain cells, e.g. when a column contains more than
one logical cell.  However, the cell coordinate retrieval code fails to
properly handle this case and will assume the index in the cell's
parent is the index in the table, which is obviously not the case for a
nested cell.

Fix this by retrieving the coordinate of the direct child containing
the given cell, whether or not it is the cell itself or one of its
parents.

Fixes #111.