Correct the cursor offset in contracted braille if the caret
is one character beyond the end of the text on the line.

Passing a cursor offset pointing beyond the end of the character buffer
results in a spurious cursor position being returned by liblouis. This
happens frequently while the user is writing text, and the caret is
positioned one character past the end of the text on the line. The
spurious cursor positions often resulted in incorrect cursor placement.

We now correct the cursor position under this condition by extending the
existing code designed to ensure that the cursor placement is
reasonable.