Don't pan the braille display to the caret position unless "pan
to cursor" is requested (i.e., panToCursor==True).

If refresh() was called with panToCursor == False and targetCursorCell
== 0, the display would be panned to a computed cursor offset anyway
(disrespecting panToCursor). In practice, this could result in a failure
to pan the display to the left with the cursor at the end of a line
(e.g., in Chromium). In this case, targetCursorCell would be set to the
caret offset on the same line, resulting in no panning being performed.