Add initial, partial support for regoin-change events
Emit the new AT-SPI2 "region-changed" signal to allow other tools
which work with Orca (e.g. screen magnifiers, highlighting tools,
etc.) to know the object and optional text range Orca is presenting.

This has so far been added for:
* Focus tracking
* Mouse Review
* Say All
* Caret navigation (all apps but web-based)

Remaining:
* Caret navigation (web-based)
* Flat review (all apps)

The above two tasks are slightly more complicated because a given
line can be made up of multiple objects being presented at once.

The current support also includes the mode/reason Orca's current
region of interest has changed. We are waiting for additional
API added to AT-SPI2 so that we can emit a signal to notify
interested tools of this mode change/reason.