Force browse mode switch in web apps when child document claims focus
If a web app gives focus directly to a child document, there's a
reasonable chance that the child document is content to be read. An
example is switching between the VSCode editor and a page containing
documentation or information. Therefore, if the document itself claims
focus and lacks attributes which suggest it should be treated as a focus
mode widget, switch to browse mode.