Handle links whose sole content is an image with alt=""
The W3C provides guidance to authors saying that they can use alt=""
on an image, including in a link, as a way to get screen readers to
ignore that image. Ideally, this ignoring would be done by the user
agent -- and is for Chrome/Chromium. But Gecko includes the useless
image in the accessibility tree, forcing screen readers to do the
pruning themselves. In addition, when a link whose sole content is
an image gets focus, Gecko emits a caret-moved event for the offset
which we're supposed to ignore. As a consequence of this event, Orca
was looking for the next presentable location. This can be a mistake
if the next presentable location happens to be another link that the
user will tab to. Therefore:

* Ignore images with alt="" that are inside links
* Don't look for the next presentable location when we get a caret-
  moved event for something which we're supposed to pretend doesn't
  exist.