Fix bug causing us to skip over objects in certain grids in browse mode
We treat ARIA gridcell objects as whole objects rather than text objects.
This, combined with an old workaround for empty text objects, was causing
us to seek the first caret context for a gridcell in the next cell. We're
now handling the empty-text problem as part of _canHaveCaretContext().
Thus this workaround can be removed.