Web: Update our command state when a braille input event occurs
When the caret moves or focus changes as a direct consequence of an
Orca command, we don't want to double-present the change. As a result,
we keep track of the nature of the last input event. Because we were
not doing this for braille input events, we sometimes had stale info
and ignored events which should have been processed and possibly
presented.