Mouse Review: Don't update current object to row to read the full row
Recently the generation of table information was cleaned up so that it
is no longer necessary to update the mouse review object to the parent
row in a table when the user wants the full row read. Furthermore,
updating the mouse review object to the row causes it to no longer be
presented because we no longer present review objects whose text consists
solely of embedded object characters.