Mouse review: Eliminate double-presentation of string in new object
When the mouse pointer is over a new object, we'll present that object's
name. For certain objects (e.g. editable text areas), we then present the
string. But for some text objects whose name comes from content, this
results in double-speaking the name. So check for this condition and
return early.