Don't automatically set locusOfFocus to newly-added web dialog
This old workaround for a Firefox bug causes us to stop preventing events
in certain web apps (e.g. Google Docs) when a dialog appears but lacks
focus. If there is a Firefox dialog which fails to give us the expected
events, we should get it fixed there or work around the problem in the
Gecko script.