Web: Fall back on object attributes for absent text attributes
Because IAccessible2 treats text attributes which apply to an entire
object (e.g. alignment) as an object attribute rather than a text
attribute, neither Gecko nor Chromium are exposing these attributes
to us via AtkText. So we need to fall back on object attributes for
those.

In addition, because this exposure is based on the CSS property name
rather than the ATK name, we need to convert the attribute name into
the expected name. Ditto for at least one value. The latter is needed
for localization.

Finally, fix a bug in Orca which was causing certain non-standard
text attribute names to always be presented, even when they had
the default value.