Add option to control automatic focus mode during native navigation
Orca already has options to automatically enter focus mode when caret
navigation and/or structural navigation result in moving to a focus-mode
widget. However, when native browser navigation (e.g. Tab) is used, we
assumed the user wanted native browser commands (e.g. typing in the
newly-focused input) to just work. Turns out we were wrong.

This commit adds a new setting to the app-specific preferences for
Gecko-based apps and Chrome/Chromium which makes it possible for users
to turn off automatically entering focus mode for native navigation.
Auto switching remains on by default so as not to change the user
experience.