Resume using AT-SPI2 cache for accessible parents in Chromium
We had to stop using the AT-SPI2 cache for parents because Chromium
was failing to emit parent-changed events causing us to not see
parts of the page and/or getting stuck. This happened, for example,
on the Github issue page with the search field.

Recently I fixed this problem in Chromium itself. Therefore we can
resume using the AT-SPI2 cache for parents.