Don't treat grids as layout-only, or gridcells as text-block elements
The heuristics we have to identify items as layout-only and text-block
assume things like grids will have a name (as per the ARIA spec). And
that the gridcells (rather than a descendant) will be what is given
focus by the author. We should instead just look at the ARIA role.