Fall back on gridcell name when generating name for nameless checkboxes
According to the ARIA spec, checkboxes are supposed to have a name.
But if the author fails to provide a name, and we're in a gridcell,
that gridcell might have presentable content. This is needed, for
example, in the Chromium history list where focus goes to descendants
of the gridcells and the checkbox for multi-selection lacks a name.