Web: Don't insist that labels for a given object must have role label
That insistence appears to be legacy code related to trying to guess
labels for badly-authored native HTML forms. And filtering out other
roles makes it harder to identify other elements serving as labels
which we should not repeat. For instance, if the user is on a heading
and then arrows down into a region with aria-labelledby pointing to
that heading, we do not want to present that name (heading) again.
Fixing that chattiness will be done in a separate commit.