Web: Cache textiness rather than the accessible text object
queryNonEmptyText() was caching the accessible text object for elements
which we concluded should be treated as navigable text. It relied upon
accessibility events to trigger the clearing of our cache. When those
events are missing, we can wind up incorrectly concluding that an element
lacks text and skip over that content. Rather than caching the text
object itself, just cache the decision about whether or not the object
should be treated as text.