Use AT-SPI2's cache for children
Not using the cache for children was an old solution to work around not
getting updates from apps/toolkits/browsers when objects were added or
removed. But not using the AT-SPI2 cache comes with a performance hit
in large accessibility trees. Therefore, we'll start using the cache for
children and address specific offenders on a more targeted basis.