Web: Ignore caret-moved events for the document itself when Tab is used
When the user Tabs into the document, we sometimes see a caret-moved
event for the document. This causes us to look for the first presentable
thing at offset and update our location to that thing. If the first
presentable thing in the document happens to be a focusable widget,
this can cause us to present the focusable widget as if it were focused
and to ignore a subsequent focus event for when it really becomes focused.
On the other hand, if the first presentable thing is a text element such
as a paragraph, we should be getting the caret moved event for that
element rather than from the document.