Fix regression in mouse clicking resulting from cache clearing
When the caret moves, we compare the object emitting the event with
the current context, searching for that context if it doesn't yet
exist. When initially clicking inside a web page, the context might
not yet exist because we cleared the cache in response to a children-
changed event. Thus when we search for the context, the result is
where we just clicked, so we say nothing. Fix this by adding an
option to not search for the context when it doesn't exist.