Web: Try to filter out duplicate caret-moved events resulting from click
We now clear our cache whenever the document or its descendant emits
children-changed events. This prevents us from potentially getting
stuck in a stale accessibility subtree. Unfortunately, if clicking
with the mouse causes the browser to emit duplicate caret-moved events
with a children-changed event in between, we'll double-speak that
caret-moved event. Try to work around this by caching the last caret
context resulting from a mouse click -- and not clearing that cache
in response to children-changed.