Web: When dumping cache due to children-changed preserve context by default
We now dump all of our cache when children-changed events are emitted.
This ensures we are not working with a stale tree (which can cause us
to get stuck and/or skip over content). Unfortunately, clearing the
context has exposed bugs in browsers. It should be safe to preserve
the context by default so we do not have to search for the current
location after a children-changed event that hasn't destroyed the
current location's accessible object.