Web: Present the current line when mouse click moves caret; set priorObj
There were special occasions (such as headings) when caret moved events
caused the entire object to be presented when clicking with the physical
mouse. This was not intentional. Move mouse-button handling to the top
of the logic where we determine what to present when the locusOfFocus
changes.

In addition, explicitly set the prior object to the old focus for mouse-
button-triggered caret-moved events. This should ensure we present roles
when initially clicking into a new element.