Web: Ensure we say something when using caret nav to interrupt SayAll
This addresses an edge case: We're in SayAll, a caret navigation command
is given which interrupts SayAll, and we happen to land in an empty text
element (e.g. a paragraph) as a consequence of that caret navigation
command. In this particular case, we should say "blank" but we weren't
doing so because we only do that "blank" fallback when we are not in
SayAll. (Interrupted SayAll is treated as in SayAll due to the fast-forward
and rewind commands.)

To address this edge case, add an optional argument to inSayAll() that
allows us to treat interrupted SayAll as not being in SayAll. Then
use that for the "blank" fallback for web content.