Web: Fix presentation of table row when caret moves due to mouse click
Don't read the full row if the last input event was a mouse button
event. We always generate the full line (i.e. each cell in the row)
when the caret moves due to a mouse click. If we think we should
read the full row for each cell in that row, we'll be extremely
chatty and also fail to present the column headers.