Web: Don't enter focus mode if we're in SayAll
Sometimes web pages auto-focus a widget after page load. If we're
doing a SayAll of that page at the time this occurs, we don't want
to leave browse mode because it can impact what and how we present
things as well as put the user in an state they don't expect.