Re-enable the "interrupt" option in speechdispatcher factory
speak() and speakUtterances() had the use of the interrupt argument
commented out back in 2007 due to unspecified issues. In 2013, we
enabled this and then immediately re-disabled it due to regressions.
Most of the code handles interruption via explicitly calling the
script's presentationInterrupt method which calls speech.stop,
which causes _cancel to be called in speechdispatcher factory. That
said, there may be downstream/external scripts/functionality that
for some reason isn't calling presentationInterrupt or stop. And
having this option in place but non functioning seems silly. So we'll
re-enable it but set the default value to False in the hopes that we
minimize the number of regressions that may be unexpectedly introduced
by this change.